Page Canonical, Internal ID, HipChat and Meta
Added logic to check for canonical based on simple / grouped ->
container product relation
Added NPS-ID to the product page
- Container products are currently showing the container product ID and
are not dynamic yet
Added HipChat module that will control notifications sent to the chat
windows
Added logic to check default meta descriptions for shortcodes and remove
them. There is a bug that is happening when a shortcode ending tag has
been removed and is not fully detected in the shortcode search which
results in the shortcode value not being removed completely. This will
need to be fixed in future updates because it effects following tags.